+++ Mall/src/main/java/com/mall/controller/PmsBrandController.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -11,12 +13,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.mall.model.PmsBrand;
-import com.mall.model.PmsProduct;
+import com.mall.model.param.PmsBrandParam;
+import com.mall.model.pms.PmsBrand;
+import com.mall.model.pms.PmsProduct;
 import com.mall.model.response.CommonResult;
-import com.mall.model.response.ProductResponse;
+import com.mall.model.response.CommonResultPmsBrand;
 import com.mall.repository.pms.PmsBrandRepository;
 
 @RequestMapping("/brand")
@@ -25,7 +29,9 @@ public class PmsBrandController {
 
 	@Autowired
 	PmsBrandRepository pmsBrandRepository;
-
+	PmsBrandRepository pmsBrandParam;
+	private static final Logger logger = LoggerFactory.getLogger(PmsProductController.class);
+	
 	// 创建品牌
 	@ResponseBody
 	@RequestMapping(value = "/create", method = RequestMethod.POST)
@@ -38,15 +44,13 @@ public class PmsBrandController {
 	}
 
 	// get商品品牌下拉栏
+	//CommonResultPmsBrand
 	@ResponseBody
 	@RequestMapping(value = "/list", method = RequestMethod.GET)
-	public CommonResult listBrand(HttpServletRequest request) {
+	public CommonResult brandList(@RequestParam PmsBrand param) {
 		// logger.info("get brand request {}", request);
-
-		Example<PmsBrand> PmsBrandExample = Example.of(pmsBrand);
-		List<pmsBrand> PmsBrandList = pmsBrandRepository.findAll(PmsBrandExample);
-		BeanUtils.copyProperties(PmsBrandList, request);
-
+		List<PmsBrand> PmsBrandList = pmsBrandParam.findAll();
+		logger.info("get brand repository {}", pmsBrandParam);
 		// TODO 把数据传到添加商品的品牌选项下拉栏
 		// logger.info("get brand repository {}", pmsBrandRepository);
 		// logger.info("get copied brand request {}", request);
@@ -55,7 +59,8 @@ public class PmsBrandController {
 //					.brandName(request.getParameter("brandName") == null ? null : String.valueOf(request.getParameter("brandName"))).build();
 //			CommonPagePmsProduct cppp = new CommonPagePmsProduct(products, 0, 0, 0l, 0);
 
-		return new CommonResult(200, request, "ok");
+		CommonResultPmsBrand commonResultPmsBrand = new CommonResultPmsBrand<>();
+		return new CommonResult(200, commonResultPmsBrand, "ok");
 	}
 
 }
