+++ Mall/src/main/java/com/mall/model/cms/CmsPrefrenceAreaProductRelation.java
@@ -1,4 +1,4 @@
-package com.mall.model;
+package com.mall.model.cms;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -7,13 +7,15 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
+import com.mall.model.pms.PmsProduct;
+
 import lombok.Data;
 
 @Entity
 @Data
-public class PmsProductFullReduction {
+public class CmsPrefrenceAreaProductRelation {
 
-	// JPA多对一
+	//JPA多对一
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "pms_product_id")
 	private PmsProduct pmsProduct;
@@ -22,10 +24,8 @@ public class PmsProductFullReduction {
 	@GeneratedValue
 	private Long id;
 
-	private double fullPrice;
+	private Long prefrenceAreaId;
 
 	private Long productId;
-
-	private double reducePrice;
 
 }
