+++ Mall/src/main/java/com/mall/model/cms/CmsSubjectProductRelation.java
@@ -1,4 +1,4 @@
-package com.mall.model;
+package com.mall.model.cms;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -6,6 +6,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+
+import com.mall.model.pms.PmsProduct;
 
 import lombok.Data;
