+++ Mall/src/main/java/com/mall/model/pms/PmsProduct.java
@@ -1,4 +1,4 @@
-package com.mall.model;
+package com.mall.model.pms;
 
 
 import java.sql.Timestamp;
@@ -10,6 +10,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
+
+import com.mall.model.cms.CmsPrefrenceAreaProductRelation;
+import com.mall.model.cms.CmsSubjectProductRelation;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
