+++ Mall/src/main/java/com/mall/model/pms/PmsProductAttributeValue.java
@@ -1,4 +1,4 @@
-package com.mall.model;
+package com.mall.model.pms;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -11,19 +11,22 @@ import lombok.Data;
 
 @Entity
 @Data
-public class CmsPrefrenceAreaProductRelation {
-
-	//JPA多对一
+public class PmsProductAttributeValue {
+	
+	// JPA多对一
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "pms_product_id")
 	private PmsProduct pmsProduct;
-
+	
 	@Id
 	@GeneratedValue
 	private Long id;
-
-	private Long prefrenceAreaId;
-
+	
+	private Long productAttributeId;
+	
 	private Long productId;
+
+	//@Column(name ="手动添加规格或参数的值，参数单值，规格有多个时以逗号隔开")
+	private String value;
 
 }
