+++ Mall/src/main/java/com/mall/model/response/CommonResultPmsBrand.java
@@ -1,0 +1,23 @@
+package com.mall.model.response;
+
+import java.util.List;
+
+import com.mall.model.pms.PmsBrand;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class CommonResultPmsBrand {
+
+	private List<PmsBrand> list;
+	private Integer pageNum;
+	private Integer pageSize;
+	private Long total;
+	private Integer totalPage;
+
+}
