+++ Mall/src/main/java/com/mall/repository/pms/PmsBrandRepository.java
@@ -2,7 +2,7 @@ package com.mall.repository.pms;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.mall.model.PmsBrand;
+import com.mall.model.pms.PmsBrand;
 
 
 public interface PmsBrandRepository extends JpaRepository<PmsBrand,Long>{
