+++ Mall/src/main/java/com/mall/repository/pms/PmsProductLadderRepository.java
@@ -2,7 +2,7 @@ package com.mall.repository.pms;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.mall.model.PmsProductLadder;
+import com.mall.model.pms.PmsProductLadder;
 
 public interface PmsProductLadderRepository extends JpaRepository<PmsProductLadder, Long>{
