+++ Mall/src/main/java/com/mall/repository/pms/PmsProductRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.mall.model.PmsProduct;
-import com.mall.model.PmsProductLadder;
+import com.mall.model.pms.PmsProduct;
+import com.mall.model.pms.PmsProductLadder;
 
 
 @Repository
