+++ Mall/src/main/java/com/mall/repository/ums/UmsMenuRepository.java
@@ -2,7 +2,7 @@ package com.mall.repository.ums;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.mall.model.UmsMenu;
+import com.mall.model.ums.UmsMenu;
 
 public interface UmsMenuRepository extends JpaRepository<UmsMenu,Long>{
