+++ Mall/src/main/java/com/mall/util/LoginUtil.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 import lombok.Data;
 import lombok.experimental.SuperBuilder;
 
-import com.mall.model.UmsMenu;
+import com.mall.model.ums.UmsMenu;
 import com.mall.repository.ums.UmsMenuRepository;
 
 @Component
